Moved the non-target binaries to separate folder
Some helper binaries (and sources) have been located in the source
folder of the firmware. These have been moved to individual folders.
elf2dfuse is a binary without sources, the "write" tool is an example
utility to use the management firmware with the host, connected to the
server to be managed.